+++ spec/PhpSpec/Console/IOSpec.php
@@ -13,7 +13,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class IOSpec extends ObjectBehavior
 {
-    function let(InputInterface $input, OutputInterface $output, DialogHelper $dialogHelper, OptionsConfig $config, Prompter $prompter)
+    function let(InputInterface $input, OutputInterface $output, OptionsConfig $config, Prompter $prompter)
     {
         $input->isInteractive()->willReturn(true);
         $input->getOption('no-code-generation')->willReturn(false);
@@ -22,7 +22,7 @@ class IOSpec extends ObjectBehavior
         $config->isCodeGenerationEnabled()->willReturn(true);
         $config->isStopOnFailureEnabled()->willReturn(false);
 
-        $this->beConstructedWith($input, $output, $dialogHelper, $config, $prompter);
+        $this->beConstructedWith($input, $output, $config, $prompter);
     }
 
     function it_has_io_interface()