+++ CHANGES.md
@@ -3,6 +3,11 @@ Upcoming
 
  * Clearer error message when specs have incorrect namespace prefix
 
+BC Breaks
+---------
+
+ * The unused `ask` and `askAndValidate` methods on `Console\IO` have been removed
+
 2.1.1 / 2015-01-09
 ==================